restructured cli commands package
