+++ cmd/skywire-cli/commands/node/pk.go
@@ -1,13 +1,15 @@
-package commands
+package node
 
 import (
 	"fmt"
+	"log"
 
+	"github.com/skycoin/skywire/cmd/skywire-cli/commands"
 	"github.com/spf13/cobra"
 )
 
 func init() {
-	rootCmd.AddCommand(pkCmd)
+	nodeCmd.AddCommand(pkCmd)
 }
 
 var pkCmd = &cobra.Command{
@@ -15,7 +17,7 @@ var pkCmd = &cobra.Command{
 	Short: "get public key of node",
 	Run: func(_ *cobra.Command, _ []string) {
 
-		client := rpcClient()
+		client := commands.PrpcClient()
 		summary, err := client.Summary()
 		if err != nil {
 			log.Fatal("Failed to connect:", err)
